Fix #2: Added in a button to navigate from the curation page, and renamed the pages folder to subfolder to hide the dead page navigations buttons at the top
